% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonIPMfit-methods.R
\name{as.array.salmonIPMfit}
\alias{as.array.salmonIPMfit}
\alias{as.matrix.salmonIPMfit}
\alias{as.data.frame.salmonIPMfit}
\title{Extract posterior samples as a matrix, array or data frame}
\usage{
\method{as.matrix}{salmonIPMfit}(object, ...)

\method{as.array}{salmonIPMfit}(object, ...)

\method{as.data.frame}{salmonIPMfit}(object, ...)
}
\arguments{
\item{object}{An object of class \link{salmonIPMfit}.}

\item{...}{Additional arguments that can be passed to \code{\link[=extract]{extract()}}.
Currently \code{pars} is the only argument supported.}
}
\value{
A \code{matrix}, \code{array} or \code{data.frame} containing the post-warmup samples of
the specified quantities.
}
\description{
Post-warmup samples from a \link{salmonIPMfit} object are converted into standard formats.
These S3 methods are wrappers for \code{as.*(object$stanfit)}; see \code{\link[=as.array.stanfit]{as.array.stanfit()}}
for details.
}
\details{
methods(class = 'stanfit')

as.array, as.data.frame, as.matrix (S3 generics: base)
extract, get_posterior_mean (S4 generics: rstan -> use extract1, stan_mean instead)
loo, loo_moment_match (S4 generics: rstan -> use S3 generics from rstanarm instead)
print
summary

methods(class = 'stanreg')

launch_shinystan (S3 generic: shinystan)
as_draws, as_draws_array, as_draws_df, as_draws_list, as_draws_matrix, as_draws_rvars (S3 generics: posterior)
log_lik
loo, loo_compare (S3 generics: loo)
posterior_predict (S3 generic: rstantools)
prior_summary (S3 generic: rstantools)
print
summary
}
\seealso{
\code{\link[=salmonIPM]{salmonIPM()}}, \link{salmonIPMfit}, \code{\link[=as.array.stanfit]{as.array.stanfit()}}
}
