% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgnorm.r
\name{dgnorm}
\alias{dgnorm}
\title{Density function for generalized normal (Subbotin) distribution}
\usage{
dgnorm(x, mu = 0, sigma = 1, shape = 2)
}
\arguments{
\item{\code{x}}{variate}

\item{\code{mu}}{mean of the distribution}

\item{\code{x}}{standard deviation of the distribution}

\item{\code{x}}{shape of the distribution}
}
\description{
Density function for a generalized normal distribution.
}

