% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgnorm_std.R
\name{dgnorm_std}
\alias{dgnorm_std}
\title{Standardized density function for generalized normal (Subbotin) distribution}
\usage{
dgnorm_std(x, mu = 0, sigma = 1, shape = 2)
}
\arguments{
\item{\code{x}}{variate}

\item{\code{mu}}{mean of the distribution}

\item{\code{x}}{standard deviation of the distribution}

\item{\code{x}}{shape of the distribution}
}
\description{
Standardized version of \code{dgnorm}.
}

