% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpexp.r
\name{dpexp}
\alias{dpexp}
\title{Density function for power-exponential distribution}
\usage{
dpexp(x, mu = 0, sigma = 1, shape = 2)
}
\arguments{
\item{x}{variate}

\item{mu}{mean of the distribution}

\item{sigma}{standard deviation of the distribution}

\item{shape}{shape of the distribution}
}
\description{
Density function for a power-exponential (also known as generalized normal
or Subbotin) distribution.
}
