% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract1}
\alias{extract1}
\title{Extract posterior samples for a single parameter}
\usage{
extract1(object, par)
}
\arguments{
\item{object}{An object of class \link{salmonIPMfit}.}

\item{par}{Character string giving quantity to return.}
}
\description{
Wrapper for \code{\link[=extract]{extract()}} applied to the \code{stanfit} component of a fitted \link{salmonIPMfit} object.
Unlike \code{\link[=extract]{extract()}}, only returns the first element, not a named list. This is useful for
obtaining the samples for a single parameter (of any dimension), though usually equivalent
to \code{as.matrix(object, par)}.
}
