% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPM_adult_sim.R
\name{IPM_adult_sim}
\alias{IPM_adult_sim}
\title{Simulate data under adult-to-adult integrated population model}
\usage{
IPM_adult_sim(pars, pop, year, X, N_age, max_age, S_H_tot, A, F_rate, B_rate,
  SR_func = "BH", n_age_tot_obs, n_HW_tot_obs)
}
\arguments{
\item{\code{pars}{Model}{parameters to b eused for simulations.}}

\item{\code{pop}{Population}{ID.}}

\item{\code{year}{}{}}

\item{\code{X}{}{}}

\item{\code{N_age}{The}{number of adult age classes.}}

\item{\code{max_age}{Oldest}{adult age class.}}

\item{\code{S_H_tot}{}{}}

\item{\code{A}{}{}}

\item{\code{F_rate}{}{}}

\item{\code{B_rate}{}{}}

\item{\code{SR_func}{Character}{code for the type of stocl-recruit model to fit. At present, only 'BH' for "Beverton-Holt' is allowed.}}

\item{\code{n_age_tot_obs}{}{}}

\item{\code{n_HW_tot_obs}{}{}}
}
\value{
A list with initial starting values for all of the parameters and states in the Stan model.
}
\description{
\code{IPM_adult_sim} simulates initial values for parameters and states in Stan.
}

