% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_defs.R
\name{par_defs}
\alias{par_defs}
\title{Generates a table of model parameters, their dimensions, and their 
definitions.}
\usage{
par_defs(stan_model = NULL)
}
\arguments{
\item{stan_model}{Character string giving the name of the Stan model being
fit (".stan" filetype extension is not included). The default argument
\code{stan_model = NULL} returns a list of all possible parameters across
all model forms.}
}
\value{
Data frame with columns for parameter name, its dimensions, and its
definition
}
\description{
Generates a table of model parameters, their dimensions, and their 
definitions.
}
