% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_model_matrix.R
\name{par_model_matrix}
\alias{par_model_matrix}
\title{Helper function to construct covariate model matrices from a list of two-sided formulas.}
\usage{
par_model_matrix(par_models, scale = TRUE, fish_data)
}
\arguments{
\item{par_models}{A list of two-sided formulas of the form
\code{theta ~ t1 + ... + tK}, where \code{theta} is a parameter or state in a \code{salmonIPM} model
that accepts covariates and \verb{t1 ... tK} are terms involving variables in
\code{fish_data}; see \code{\link[=salmonIPM]{salmonIPM()}} for details.}

\item{scale}{Logical indicating whether the terms in model matrices
constructed from \code{fish_data} using the formulas in \code{par_models} should be scaled to have
column SDs of 1 in addition to being centered (\code{TRUE}) or centered only (\code{FALSE}).}

\item{fish_data}{See \code{\link[=salmonIPM]{salmonIPM()}}. In particular, the columns \code{...} may be used on
the right-hand side of formulas in \code{par_models}.}
}
\value{
A named list of model matrices corresponding to the elements of \code{par_models}.
}
\description{
Helper function to construct covariate model matrices from a list of two-sided formulas.
}
\seealso{
\code{\link[=salmonIPM]{salmonIPM()}} for specifying covariate effects via \code{par_models},
\code{\link[=stan_data]{stan_data()}} for assembling input data including covariate model matrices
}
