% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\alias{normal}
\alias{gnormal}
\alias{lognormal}
\alias{beta}
\alias{dirichlet}
\title{Prior distributions}
\usage{
normal(mean = 0, sd = 1)

gnormal(mean = 0, scale = 1, shape = 1)

lognormal(meanlog = 0, sdlog = 1)

beta(a = 1, b = 1)

dirichlet(concentration = 1)
}
\arguments{
\item{mean}{Prior mean for normal or generalized normal distribution.}

\item{sd}{Prior standard deviation for normal distribution.}

\item{scale}{Prior scale for generalized normal distribution. Equivalent to
\code{alpha} in \link[gnorm:gnorm]{gnorm::gnorm}, but renamed to avoid confusion with the spawner-recruit
intrinsic productivity parameter.}

\item{shape}{Prior shape for generalized normal distribution. Equivalent to \code{beta}
in \link[gnorm:gnorm]{gnorm::gnorm}.}

\item{meanlog, sdlog}{Prior log-scale mean and standard deviation, respectively, for
lognormal distribution; see \link[stats:Lognormal]{stats::Lognormal}.}

\item{a, b}{Prior shape parameters for the Beta distribution. Equivalent to
\code{shape1} and \code{shape2}, respectively, in \link[stats:Beta]{stats::Beta}.}

\item{concentration}{Vector of shape parameters for the Dirichlet distribution.
Equivalent to \code{alpha} in \link[gtools:dirichlet]{gtools::dirichlet}, but renamed to avoid confusion with
the spawner-recruit intrinsic productivity parameter.}
}
\value{
A named list to be used internally by the \strong{salmonIPM} model-fitting functions.
}
\description{
The functions described on this page are used to specify priors
on selected (hyper)parameters in \strong{salmonIPM} models.

The default priors used in the various models
are intended to be weakly informative in that they provide moderate
regularization and help stabilize computation. Priors on scaling parameters, e.g.
\code{Rmax} or \code{mu_Rmax}, are automatically adjusted to be weakly informative but
consistent with overall population density. For many applications these
defaults will perform well, but if external information not included in
\code{fish_data} is available, it can be incorporated via user-specified priors on
key parameters. See \strong{Details} for a table of available prior options.
}
\details{
The table below shows the parameters in each model that can be given
user-specified priors and the corresponding distributions. Note that
users can modify the prior parameters but not the distribution families; attempting
to do the latter will result in an error.

The generalized normal density with \verb{shape >> 1} is useful as a "soft-uniform" prior
to regularize the posterior away from regions of parameter space that may cause
computational or sampling problems. In the case of spawner and smolt observation error
log-SDs, the default prior bounds them ≳ 0.1.

For parameters that are modeled as functions of covariates (see \code{\link[=salmonIPM]{salmonIPM()}}),
the specified prior applies when all predictors are at their sample means.\tabular{lccccccccccccc}{
    \tab  \tab  \tab  \tab  \tab  \tab  \tab \strong{Parameter} \tab  \tab  \tab  \tab  \tab  \tab  \cr
   \strong{Model} \tab \strong{\code{alpha}} \tab \strong{\code{mu_alpha}} \tab \strong{\code{mu_psi}} \tab \strong{\code{Rmax}} \tab \strong{\code{mu_Rmax}} \tab \strong{\code{Mmax}} \tab \strong{\code{mu_Mmax}} \tab \strong{\code{mu_MS}} \tab \strong{mu_p} \tab \strong{\code{mu_SS}} \tab \strong{\code{tau}} \tab \strong{\code{tau_S}} \tab \strong{\code{tau_M}} \cr
   \code{IPM_SS_np} \tab lognormal \tab  \tab  \tab lognormal \tab  \tab  \tab  \tab  \tab dirichlet \tab  \tab gnormal \tab  \tab  \cr
   \code{IPM_SSiter_np} \tab lognormal \tab  \tab  \tab lognormal \tab  \tab  \tab  \tab  \tab dirichlet \tab beta \tab gnormal \tab  \tab  \cr
   \code{IPM_SS_pp} \tab  \tab normal \tab  \tab  \tab normal \tab  \tab  \tab  \tab dirichlet \tab  \tab gnormal \tab  \tab  \cr
   \code{IPM_SSiter_pp} \tab  \tab normal \tab  \tab  \tab normal \tab  \tab  \tab  \tab dirichlet \tab beta \tab gnormal \tab  \tab  \cr
   \code{IPM_SMS_np} \tab lognormal \tab  \tab  \tab  \tab  \tab lognormal \tab  \tab beta \tab dirichlet \tab  \tab  \tab gnormal \tab gnormal \cr
   \code{IPM_SMS_pp} \tab  \tab normal \tab  \tab  \tab  \tab  \tab normal \tab beta \tab dirichlet \tab  \tab  \tab gnormal \tab gnormal \cr
   \code{IPM_SMaS_np} \tab lognormal \tab  \tab  \tab  \tab  \tab lognormal \tab  \tab  \tab  \tab  \tab  \tab gnormal \tab gnormal \cr
   \code{IPM_ICchinook_pp} \tab  \tab  \tab  \tab  \tab  \tab  \tab normal \tab  \tab dirichlet \tab  \tab  \tab gnormal \tab gnormal \cr
   \code{IPM_LCRchum_pp} \tab  \tab  \tab beta \tab  \tab  \tab  \tab normal \tab beta \tab dirichlet \tab  \tab  \tab gnormal \tab gnormal \cr
}
}
