% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonIPM.R
\name{salmonIPM}
\alias{salmonIPM}
\title{Fits an integrated or run-reconstruction spawner-recruit model.}
\usage{
salmonIPM(fish_data, env_data = NULL, model, pool_pops = TRUE,
  init = NULL, pars = NULL, chains, iter, warmup, thin = 1, cores = 3,
  control = NULL)
}
\arguments{
\item{fish_data}{Data frame that includes the following \code{colnames}, in no particular order except where noted:
\describe{
\item{\code{pop}}{Numeric or character population ID.}
\item{\code{year}}{Numeric variable giving the year the fish spawned (i.e., the brood year).}
\item{\code{A}}{Spawning habitat size (either stream length or area). Will usually be time-invariant within a population, but need not be.}
\item{\code{S_tot_obs}}{Total number (not density) of wild and hatchery-origin spawners.}
\item{\code{n_age_minAge...n_age_maxAge}}{Multiple columns of observed spawner age frequencies (i.e., counts), where minAge (maxAge) is the numeral age in years (total, not ocean age) of the youngest (oldest) spawners.}
\item{\code{n_W_obs}}{Observed frequency of natural-origin spawners.}
\item{\code{n_H_obs}}{Observed frequency of hatchery-origin spawners.}
\item{\code{fit_p_HOS}}{Logical or 0/1 indicating for each row in fish_data whether the model should estimate p_HOS > 0. This is only required if model == "IPM".}
\item{\code{F_rate}}{Total harvest rate (proportion) of natural-origin fish.}
\item{\code{B_take_obs}}{Number of adults taken for hatchery broodstock.}
}}

\item{env_data}{Optional data frame whose variables are time-varying environmental covariates, sequentially ordered with each row corresponding to a unique year in fish_data.}

\item{model}{One of \code{"IPM"} or \code{"RR"}, indicating whether the data are intended for an integrated or run-reconstruction model.}

\item{pool_pops}{Logical, with default \code{TRUE}, indicating whether or not to treat the different populations as hierarchical rather than fixed/independent.}

\item{init}{A list of named lists of initial values to be passed to \code{stan}. If \code{NULL}, initial values will be automatically generated from the supplied data using \code{stan_init}.}

\item{pars}{A vector of character strings specifying parameters to monitor. If NULL, default values are used. If a non-default value is supplied, it is the user's responsibility to make sure the parameters requested appear in the model configuration specified.}

\item{chains}{A positive integer specifying the number of Markov chains.}

\item{iter}{A positive integer specifying the number of iterations for each chain (including warmup).}

\item{warmup}{A positive integer specifying the number of warmup (aka burnin) iterations per chain. If step-size adaptation is on (which it is by default), this also controls the number of iterations for which adaptation is run (and hence these warmup samples should not be used for inference). The number of warmup iterations should not be larger than \code{iter}}

\item{thin}{A positive integer specifying the period for saving samples. The default is 1, which is usually the recommended value.}

\item{cores}{Number of cores to use when executing the chains in parallel. Defaults to 3.}

\item{control}{A named list of parameters to control the sampler's behavior (see \code{rstan::stan} for details). It defaults to NULL so all the default values are used.}
}
\value{
An object of class \code{stanfit} representing the fitted model. See \code{rstan::stan} for details.
}
\description{
Fits an integrated or run-reconstruction spawner-recruit model.
}
