% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonIPMfit-class.R
\name{salmonIPMfit-class}
\alias{salmonIPMfit-class}
\alias{salmonIPMfit}
\title{Fitted model objects: the \code{salmonIPMfit} class}
\usage{
salmonIPMfit(
  stanfit,
  call,
  model,
  life_cycle,
  pool_pops,
  SR_fun,
  RRS,
  par_models,
  center,
  scale,
  prior,
  age_S_obs,
  age_S_eff,
  conditionGRonMS
)
}
\arguments{
\item{stanfit}{An object of class \linkS4class{stanfit}.}

\item{call}{An object of class \linkS4class{call} containing the call to \code{\link[=salmonIPM]{salmonIPM()}}.}

\item{model}{Either \code{"IPM"} or \code{"RR"}, indicating whether to fit an IPM
or run-reconstruction regression model.}

\item{life_cycle}{Character string indicating which life-cycle model to fit.
One of the following options (must be \code{"SS"} if \code{model == "RR"}):
\itemize{
\item \code{"SS"}  Spawner-to-spawner (the default)
\item \code{"SSiter"} Spawner-to-spawner with iteroparity (an alias for \code{"SS"} with
\code{stan_data("IPM_SS_[x]p", fish_data = fish_data)$iter} set to 1)
\item \code{"SMS"}  Spawner-smolt-spawner with a fixed smolt age
\item \code{"SMaS"}  Spawner-smolt-spawner with multiple smolt age classes (currently only
available for \code{pool_pops == FALSE})
\item \code{"LCRchum"}  Customized spawner-smolt-spawner model for Lower Columbia River chum
(\code{pool_pops == TRUE})
}}

\item{pool_pops}{Logical defaulting to \code{TRUE}, indicating whether
to model multiple populations hierarchically rather than as independent "fixed effects".
It is possible to fit a model to multiple populations simultaneously even though they
share no parameters; indeed this is more efficient than fitting them
one at a time because calculations are vectorized and warmup is shared.}

\item{SR_fun}{One of \code{"exp"} (density-independent discrete exponential),
\code{"BH"} (Beverton-Holt, the default), or \code{"Ricker"}, indicating which spawner-recruit
function to fit. Synonyms \code{"DI"}, \code{"B-H"}, \code{"bh"}, \code{"b-h"} and \code{"ricker"} are accepted.}

\item{RRS}{A character string or vector of strings naming parameters of the function specified by
\code{SR_fun} that differ between wild- and hatchery-origin spawners, such that the
relative reproductive success of hatchery spawners is not equal to 1. If \code{pool_pops == TRUE},
these should be the names of the population-specific parameters, not their hyper-means.
For example, if \code{life_cycle \%in\% c("SS","SSiter")}, the options are
\code{"none"} (the default), \code{"alpha"}, \code{"Rmax"}, or \code{c("alpha","Rmax")}.
Currently \code{RRS} is only implemented for \code{pool_pops == FALSE}.}

\item{par_models}{Optional list of two-sided formulas of the form
\code{theta ~ x1 + ... + xK}, where \code{theta} is a (hyper)parameter or state in the model
specified by \code{stan_model} that accepts covariates (see \strong{Details} for available
model-parameter combinations) and \code{x1 + ... + xK} are terms involving variables in
\code{fish_data}. Standard formula syntax such as \code{:} and \code{*} may be used;
see \code{\link[stats:formula]{stats::formula()}}.}

\item{center}{Logical indicating whether the terms in model matrices
constructed from \code{fish_data} using the formulas in \code{par_models} should be centered.
It is usually recommended to use the default (\code{TRUE}) so the baseline parameter
estimate applies when predictors are at their sample means, but in some cases such
as factor predictors \code{center = FALSE} may be appropriate. If combining categorical
and numeric predictors, the latter can be centered and scaled prior to modeling.}

\item{scale}{Logical indicating whether the model matrices constructed from
\code{fish_data} using the formulas in \code{par_models} should be scaled to have
column SDs of 1. Unit-scaling predictors is less critical than centering, but is advisable
if variables have scales far from 1.}

\item{prior}{Optional list of two-sided formulas of the form
\code{theta ~ distribution(params)}, where \code{theta} is a hyperparameter that can take
a user-specified prior and \code{distribution()} is its canonical prior family. See
\code{\link{priors}} for details on the available parameters in each model and their
corresponding families. Any hyperparameters not given explicit priors
will use the default values of the prior \code{params}.}

\item{age_S_obs}{If \code{stan_model == "IPM_SS_pp"}, a logical or 0/1 integer
vector indicating, for each adult age, whether the observed total spawner data
includes that age. The default is to treat \code{S_obs} as including spawners of
all ages. This option may be useful if certain age classes are not counted. If \code{life_cycle == "SSiter"}, \code{N_age} refers to the total
number of maiden and repeat age classes (counting the repeat plus group as 1).}

\item{age_S_eff}{If \code{stan_model == "IPM_SS_pp"}, a logical or 0/1
vector indicating, for each adult age, whether spawners of that age
contribute to reproduction. This can be used, e.g., to exclude jacks
from the effective breeding population. The default is to include spawners
of all ages. If \code{life_cycle == "SSiter"}, \code{N_age} refers to the total
number of maiden and repeat age classes (counting the repeat plus group as 1).}

\item{conditionGRonMS}{If \code{life_cycle == "SMaS"}, logical indicating
whether the Gilbert-Rich age frequencies \code{n_GRage_obs} in \code{fish_data} are conditioned
on ocean age. If \code{FALSE} (the default) the counts are assumed to be sampled randomly
from the population. If \code{TRUE}, it is assumed that the number of spawners of each ocean
age (e.g., jacks vs 1-ocean) is arbitrary, but smolt (FW) age is randomly sampled within
each ocean age; i.e., in a \verb{smolt age x ocean age} contingency table, the cell frequencies
are conditioned on the column totals.}
}
\value{
A \code{salmonIPMfit} object.
}
\description{
Models fitted with \link{salmonIPM-package} are objects of S3 class \code{salmonIPMfit}. Elements
generally mirror the named arguments passed to the constructor function by \code{\link[=salmonIPM]{salmonIPM()}}
and are mostly self-explanatory, except \code{prior.info} is a named list whose elements are
lists returned by \link{priors} functions for all hyperparameters in the model, whether
modifiable (user-specified or default) or hard-coded.
}
\seealso{
\code{\link[=salmonIPM]{salmonIPM()}}, \link{priors}
}
