% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simIPM.R
\name{simIPM}
\alias{simIPM}
\title{Simulate data from a salmonid integrated population model}
\usage{
simIPM(
  life_cycle = "SS",
  SR_fun = "BH",
  pars,
  par_models = NULL,
  scale = TRUE,
  N_age,
  max_age,
  ages = NULL,
  age_F = NULL,
  age_B = NULL,
  fish_data
)
}
\arguments{
\item{life_cycle}{Character string indicating which life-cycle model to
simulate. Currently available options are spawner-to-spawner (\code{"SS"}, the default),
iteroparous spawner-to-spawner (\code{"SSiter"}), or spawner-smolt-spawner (\code{"SMS"}).}

\item{SR_fun}{One of \code{"exp"} (density-independent discrete exponential),
\code{"BH"} (Beverton-Holt, the default), or \code{"Ricker"}, indicating which spawner-recruit
function to fit. Synonyms \code{"DI"}, \code{"B-H"}, \code{"bh"}, \code{"b-h"} and \code{"ricker"} are accepted.}

\item{pars}{Named list of (hyper)parameters to be used for simulations:
\itemize{
\item \code{mu_alpha}  Hyper-mean of log intrinsic productivity.
\item \code{beta_alpha}  Vector of regression coefficients for log intrinsic productivity.
Must be specified if \code{par_models} includes \code{alpha ~ ...}; otherwise will be ignored
and may be omitted.
\item \code{sigma_alpha}  Hyper-SD of log intrinsic productivity.
\item \code{mu_Rmax}  If \code{life_cycle == "SS"}, hyper-mean of log maximum recruitment.
\item \code{beta_Rmax}  If \code{life_cycle == "SS"}, vector of regression coefficients for
log maximum recruitment. Must be specified if \code{par_models} includes \code{Rmax ~ ...};
otherwise will be ignored and may be omitted.
\item \code{sigma_Rmax}  If \code{life_cycle == "SS"}, hyper-SD of log maximum recruitment.
\item \code{rho_alphaRmax}  If \code{life_cycle == "SS"}, correlation between log(alpha) and log(Rmax).
\item \code{beta_R}  If \code{life_cycle == "SS"}, vector of regression coefficients for
log recruitment. Must be specified if \code{par_models} includes \code{R ~ ...};
otherwise will be ignored and may be omitted.
\item \code{rho_R}  If \code{life_cycle == "SS"}, AR(1) coefficient of brood-year
log productivity anomalies.
\item \code{sigma_year_R}  If \code{life_cycle == "SS"}, hyper-SD of brood-year
log productivity anomalies.
\item \code{sigma_R}  If \code{life_cycle == "SS"}, unique recruitment process error SD.
\item \code{mu_Mmax}  If \code{life_cycle == "SMS"}, hyper-mean of log maximum smolt recruitment.
\item \code{beta_Mmax}  If \code{life_cycle == "SMS"}, vector of regression coefficients for
log maximum smolt recruitment.
\item \code{sigma_Mmax}  If \code{life_cycle == "SMS"}, hyper-SD of log maximum smolt recruitment.
\item \code{rho_alphaMmax}  If \code{life_cycle == "SMS"}, correlation between log(alpha) and log(Mmax).
\item \code{beta_M}  If \code{life_cycle == "SMS"}, vector of regression coefficients for
log smolt recruitment. Must be specified if \code{par_models} includes \code{M ~ ...};
otherwise will be ignored and may be omitted.
\item \code{rho_M}  If \code{life_cycle == "SMS"}, AR(1) coefficient of spawner-smolt log
productivity anomalies.
\item \code{sigma_year_M}  If \code{life_cycle == "SMS"}, process error SD
of spawner-smolt log productivity anomalies.
\item \code{sigma_M}  If \code{life_cycle == "SMS"},
SD of unique spawner-smolt productivity process errors.
\item \code{mu_MS}  If \code{life_cycle == "SMS"}, mean SAR.
\item \code{beta_MS}  If \code{life_cycle == "SMS"}, vector of regression coefficients for
logit SAR anomalies. Must be specified if \code{par_models} includes \code{s_MS ~ ...};
otherwise will be ignored and may be omitted.
\item \code{rho_MS}  If \code{life_cycle == "SMS"}, AR(1) coefficient for logit SAR anomalies.
\item \code{sigma_year_MS}  If \code{life_cycle == "SMS"}, process error SD of logit SAR anomalies.
\item \code{sigma_MS}  If \code{life_cycle == "SMS"}, SD of unique logit SAR process errors.
\item \code{mu_p}  Among-population mean simplex of age distributions.
\item \code{sigma_pop_p}  Vector of among-population SDs of mean log-ratio age distributions.
\item \code{R_pop_p}  Among-population correlation matrix of mean log-ratio age distributions.
\item \code{sigma_pop_p}  Vector of SDs of log-ratio cohort age distributions.
\item \code{R_pop_p}  Correlation matrix of cohort log-ratio age distributions.
\item \code{sigma_p}  Vector of SDs of log-ratio cohort age distributions.
\item \code{R_p}  Correlation matrix of cohort log-ratio age distributions.
\item \code{mu_SS}  If \code{life_cycle == "SSiter"}, mean kelt survival.
\item \code{beta_SS}  If \code{life_cycle == "SSiter"}, vector of regression coefficients for
logit kelt survival. Must be specified if \code{par_models} includes \code{s_SS ~ ...};
otherwise will be ignored and may be omitted.
\item \code{rho_SS}  If \code{life_cycle == "SSiter"}, AR(1) coefficient for annual logit kelt survival anomalies.
\item \code{sigma_year_SS}  If \code{life_cycle == "SSiter"}, process error SD of annual logit kelt survival anomalies.
\item \code{sigma_SS}  If \code{life_cycle == "SSiter"}, SD of unique logit kelt survival process errors.
\item \code{tau}  If \code{life_cycle != "SMS"}, spawner observation error SD.
\item \code{tau_M}  If \code{life_cycle == "SMS"}, smolt observation error SD.
\item \code{tau_S}  If \code{life_cycle == "SMS"}, pawner observation error SD.
\item \code{S_init_K}  Mean of initial spawning population size as a fraction of carrying capacity.
}}

\item{par_models}{Optional list of two-sided formulas of the form
\code{theta ~ x1 + ... + xK}, where \code{theta} is a (hyper)parameter or state in the model
specified by \code{stan_model} that accepts covariates (see \strong{Details} for available
model-parameter combinations) and \code{x1 + ... + xK} are terms involving variables in
\code{fish_data}. Standard formula syntax such as \code{:} and \code{*} may be used;
see \code{\link[stats:formula]{stats::formula()}}.}

\item{scale}{Logical indicating whether the model matrices constructed from
\code{fish_data} using the formulas in \code{par_models} should be scaled to have
column SDs of 1. Unit-scaling predictors is less critical than centering, but is advisable
if variables have scales far from 1.}

\item{N_age}{Number of (maiden) adult age classes.}

\item{max_age}{Oldest (maiden) adult age class.}

\item{ages}{For multi-stage models, a named list giving the fixed ages
in years of all subadult life stages. For example, if \code{life_cycle == "SMS"},
\code{list(M = a)} where \code{a} is integer smolt age.}

\item{age_F}{Logical or 0/1 vector of length \code{N_age} indicating
whether each adult age is fully (non)selected by the fishery.
The default is all selected. If \code{life_cycle == "SSiter"}, \code{N_age} refers to the total
number of maiden and repeat age classes (counting the repeat plus group as 1).}

\item{age_B}{Logical or 0/1 vector of length \code{N_age} indicating
whether each adult age is fully (non)selected in broodstock collection.
The default is all selected. If \code{life_cycle == "SSiter"}, \code{N_age} refers to the total
number of maiden and repeat age classes (counting the repeat plus group as 1).}

\item{fish_data}{Data frame with columns:
\itemize{
\item \code{pop}  Numeric, character or factor population ID.
\item \code{year}  Numeric or integer giving the year the fish spawned (i.e., the brood year).
\item \code{A}  Spawning habitat size (either stream length or area). Will often be
time-invariant within a population, but need not be.
\item \code{p_HOS}  True fraction of hatchery-origin spawners.
\item \code{F_rate}  Total harvest rate (proportion) of natural-origin fish.
\item \code{B_rate}  Adult hatchery broodstock removal rate.
\item \code{n_age_obs}  Number of spawners sampled for age composition.
\item \code{n_HW_obs}  Number of spawners sampled for hatchery / wild origin.
\item \code{...}  Additional variables to be used as covariates.
}}
}
\value{
A named list with elements
\itemize{
\item \code{sim_dat}  A data frame containing simulated data in the structure of
(see \code{\link[=salmonIPM]{salmonIPM()}}) appropriate for the specified \code{life_cycle}, ready to be passed
to \code{\link[=salmonIPM]{salmonIPM()}}.
\item \code{pars_out}  A named list of hyperparameters, group-level parameters, and states
used in generating the pseudo-data.
}
}
\description{
Generate pseudo-data, group-level parameters, and states from a specified
\strong{salmonIPM} integrated population model given values for the hyperparameters.
This may be useful, e.g., in simulation-based calibration and model sensitivity
checking.
}
\seealso{
\code{\link[=salmonIPM]{salmonIPM()}} for fitting models
}
