% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_mean.r
\name{stan_mean}
\alias{stan_mean}
\title{Extract posterior means from Stan object}
\usage{
stan_mean(object, pars)
}
\arguments{
\item{object}{A fitted Stan object.}

\item{pars}{Character vector with the names of parameters to summarize.}
}
\value{
A scalar or vector of means of the posterior distribution of \code{pars}.
}
\description{
Convenience function to simplify extracting posterior means or single
parameters from stanfit objects.
}
