% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_pars.R
\name{stan_pars}
\alias{stan_pars}
\title{Helper function that returns the parameters and states that a given
model should save.}
\usage{
stan_pars(stan_model)
}
\arguments{
\item{stan_model}{Character string giving the name of the Stan model being
fit (\code{".stan"} filetype extension is not included).}
}
\value{
Character vector with names of parameters and states that Stan will
save.
}
\description{
Helper function that returns the parameters and states that a given
model should save.
}
