% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_prior.R
\name{stan_prior}
\alias{stan_prior}
\title{Validate prior specification for an integrated spawner-recruit model.}
\usage{
stan_prior(value, default)
}
\arguments{
\item{value}{A user-specified prior function as described in \code{\link{priors}}.}

\item{default}{The default prior function as described in \code{\link{priors}}.}
}
\value{
A named vector of user-specifiable prior parameters whose elements are part of the
\code{data} argument passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} when fitting \strong{salmonIPM} models.
}
\description{
Takes an optional user-specified prior function and otherwise returns default
prior parameters for the given \code{stan_model}. This function is for internal use.
}
