% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_unicode.R
\name{test_unicode}
\alias{test_unicode}
\title{Test if unicode is rendering in test_unicode.Rd}
\usage{
test_unicode()
}
\description{
Test if unicode is rendering in test_unicode.Rd
}
\details{
This should be a checkbox: ☑
But maybe the backslash escape version will work: \\u2611;
}
