% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract1.r
\name{extract1}
\alias{extract1}
\title{Extracts fitted Stan object}
\usage{
extract1(object, par)
}
\arguments{
\item{object}{fitted Stan object}

\item{par}{parameters/states to return}
}
\description{
Improved function for extracting fitted Stan objects.
}
