% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPM_adult_sim.R
\name{IPM_adult_sim}
\alias{IPM_adult_sim}
\title{Simulate data under adult-to-adult integrated population model}
\usage{
IPM_adult_sim(pars, pop, year, X = NULL, N_age, max_age, A, F_rate, B_rate,
  SR_func = "BH", n_age_tot_obs, n_HW_tot_obs)
}
\arguments{
\item{pars}{Model parameters to be used for simulations.}

\item{pop}{Population ID.}

\item{year}{Calendar year.}

\item{X}{Covariate(s).}

\item{N_age}{The number of adult age classes.}

\item{max_age}{Oldest adult age class.}

\item{A}{Area of spawning habitat.}

\item{F_rate}{Harvest rate.}

\item{B_rate}{Broodstock take rate.}

\item{SR_func}{Character code for the type of stocl-recruit model to fit. At present, only 'BH' for "Beverton-Holt' is allowed.}

\item{n_age_tot_obs}{Number of adults for age comp.}

\item{n_HW_tot_obs}{Number of adults for H vs W origin.}
}
\value{
A list with initial starting values for all of the parameters and states in the Stan model.
}
\description{
\code{IPM_adult_sim} simulates initial values for parameters and states in Stan.
}
