% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_defs.R
\name{par_defs}
\alias{par_defs}
\title{Returns a list of the parameters returned by default for the specified
model.}
\usage{
par_defs(stan_model = NULL)
}
\arguments{
\item{stan_model}{One of two options:
\enumerate{
  \item A \code{character} string giving the name of the Stan model being fit
  (without the ".stan" filetype extension)
  \item An object of class \code{stanfit} as returned by a call to
  \code{\link{salmonIPM}}:
}

The default argument \code{stan_model = NULL} returns a list of all of the 
parameters returned by default, across all model forms.}
}
\value{
Data frame with columns for parameter name, its dimensions, and its
definition
}
\description{
Returns a list of the parameters returned by default for the specified
model.
}
