% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_defs.R
\name{par_defs}
\alias{par_defs}
\title{Returns a table of the parameters and their descriptions, as returned by
default for the specified model.}
\usage{
par_defs(stan_model = NULL)
}
\arguments{
\item{stan_model}{Either a character string giving the name of the Stan model being fit
(without the \code{".stan"} filetype extension) or an object of class \code{stanfit}
as returned by a call to \code{\link[=salmonIPM]{salmonIPM()}}. The default value returns a list of
all parameters returned by default, across all model forms.}
}
\value{
Data frame with 3 columns listing the parameters, their dimensions,
and their definitions.
}
\description{
Returns a table of the parameters and their descriptions, as returned by
default for the specified model.
}
