% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{priors}
\alias{priors}
\alias{normal}
\alias{gnormal}
\alias{lognormal}
\alias{beta}
\alias{dirichlet}
\title{Prior distributions}
\usage{
normal(mean = 0, sd = 1)

gnormal(mean = 0, scale = 1, shape = 1)

lognormal(meanlog = 0, sdlog = 1)

beta(a = 1, b = 1)

dirichlet(concentration = 1)
}
\arguments{
\item{mean}{Prior mean for normal or generalized normal distribution.}

\item{sd}{Prior standard deviation for normal distribution.}

\item{scale}{Prior scale for generalized normal distribution. Equivalent to
\code{alpha} in \code{\link[gnorm:gnorm]{gnorm::gnorm}}, but renamed to avoid confusion with the spawner-recruit
intrinsic productivity parameter.}

\item{shape}{Prior shape for generalized normal distribution. Equivalent to \code{beta}
in \code{\link[gnorm:gnorm]{gnorm::gnorm}}.}

\item{meanlog, sdlog}{Prior log-scale mean and standard deviation, respectively, for
lognormal distribution; see \code{\link[stats:Lognormal]{stats::Lognormal}}.}

\item{a, b}{Prior shape parameters for the Beta distribution. Equivalent to
\code{shape1} and \code{shape2}, respectively, in \code{\link[stats:Beta]{stats::Beta}}.}

\item{concentration}{Vector of shape parameters for the Dirichlet distribution.
Equivalent to \code{alpha} in \code{\link[gtools:dirichlet]{gtools::dirichlet}}, but renamed to avoid confusion with
the spawner-recruit intrinsic productivity parameter.}
}
\value{
A named list to be used internally by the \strong{salmonIPM} model-fitting functions.
}
\description{
These functions are used to specify priors
on selected (hyper)parameters in \strong{salmonIPM} models.

The default priors used in the various models
are intended to be weakly informative, in that they provide moderate
regularization and help stabilize computation. Priors on scaling parameters, e.g.
\code{Rmax} or \code{mu_Rmax}, are automatically adjusted to be weakly informative but
consistent with overall population density. For many applications these
defaults will perform well, but if external information not included in
\code{fish_data} is available, it can be incorporated via user-specified priors on
key parameters. See \strong{Details} for a table of available prior options.
}
\details{
The table below shows the parameters in each model that can be given
user-specified priors and the corresponding distributions. Note that
users can modify the prior parameters but not the distribution families; attempting
to do the latter will result in an error.

Priors for parameters that are bounded on the positive real line
(e.g. \code{tau}, \code{tau_S} and \code{tau_M}) are automatically left-truncated at zero.

For parameters that are modeled as functions of covariates (see \code{\link[=salmonIPM]{salmonIPM()}}),
the specified prior applies when all predictors are at their sample means.

If \code{RRS != "none"}, the global spawner-recruit parameters must be replaced with their
\code{W} and \code{H} counterparts; e.g. if \code{RRS == "alpha"} then instead of a prior on \code{alpha}
one would specify priors on \code{alpha_W} and \code{alpha_H}. If the former is provided, it will
have no effect. See \code{\link[=salmonIPM]{salmonIPM()}} for details of the \code{RRS} argument.

The generalized normal density with \verb{shape >> 1} is useful as a platykurtic "soft-uniform"
prior to regularize the posterior away from regions of parameter space that may cause
computational or sampling problems. In the case of spawner and smolt observation error
log-SDs, the default prior bounds them ≳ 0.1.\tabular{lcccccccccccc}{
    \tab  \tab  \tab  \tab  \tab  \tab  \tab \strong{Parameter (PDF)} \tab  \tab  \tab  \tab  \tab  \cr
   \strong{Model} \tab \code{alpha} \cr \code{lognormal} \tab \code{mu_alpha} \cr \code{normal} \tab \code{mu_psi} \cr \code{beta} \tab \code{Rmax} \cr \code{lognormal} \tab \code{mu_Rmax} \cr \code{normal} \tab \code{Mmax} \cr \code{lognormal} \tab \code{mu_Mmax} \cr \code{normal} \tab \code{mu_MS} \cr \code{beta} \tab \code{mu_p} \cr \code{dirichlet} \tab \code{mu_SS} \cr \code{beta} \tab \code{tau} \cr \code{gnormal} \tab \code{tau_S} \cr \code{tau_M} \cr \code{gnormal} \cr
   \code{IPM_SS_np} \tab ☑ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☑ \tab ☐ \cr
   \code{IPM_SSiter_np} \tab ☑ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☑ \tab ☑ \tab ☐ \cr
   \code{IPM_SS_pp} \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☑ \tab ☐ \cr
   \code{IPM_SSiter_pp} \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☑ \tab ☑ \tab ☐ \cr
   \code{IPM_SMS_np} \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☑ \tab ☑ \tab ☐ \tab ☐ \tab ☑ \cr
   \code{IPM_SMS_pp} \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☑ \tab ☑ \tab ☐ \tab ☐ \tab ☑ \cr
   \code{IPM_SMaS_np} \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \cr
   \code{IPM_ICchinook_pp} \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☑ \cr
   \code{IPM_LCRchum_pp} \tab ☐ \tab ☐ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \tab ☑ \tab ☑ \tab ☑ \tab ☐ \tab ☐ \tab ☐ \cr
}
}
