% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_recon.r
\name{run_recon}
\alias{run_recon}
\title{Standardized density function for generalized normal (Subbotin) distribution}
\usage{
run_recon(data)
}
\arguments{
\item{data}{data frame with the following \code{colnames}:
\describe{
\item{\code{brood.yr}}{The year the fish were spawned.}
\item{\code{pop}}{Population ID.}
\item{\code{p3...pN}}{Multiple columns of the proportion of age-3, age-4, ..., age-N fish.}
\item{\code{pHOS}}{The proportion of Hatchery Origin Spawners.}
\item{\code{nS}}{The total number of wild + hatchery-origin spawners.}
\item{\code{wild.broodstk}}{The number of fish taken for hatchery broodstock.}
\item{\code{hrate.w}}{The harvest rate (0-1) of wild fish.}
}}
}
\value{
A list with the following:
\describe{
\item{\code{R}}{A [years x ages] matrix of age-specific recruits by brood year.}
\item{\code{R_tot}}{A vector with the total number of estimated of recruits across all ages for each brood year.}
}
}
\description{
Standardized version of \code{dgnorm}.
}
