% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_recon.r
\name{run_recon}
\alias{run_recon}
\title{Perform run reconstruction on brood table data}
\usage{
run_recon(fish_data, age_F = NULL, age_B = NULL)
}
\arguments{
\item{fish_data}{Data frame that includes the following columns,
in no particular order except where noted:
\itemize{
\item \code{pop}  Numeric, factor or character population ID.
\item \code{year}  Numeric variable giving the year the fish spawned (i.e., the brood year).
\item \code{A} Spawning habitat size (either stream length or area).
Will often be time-invariant within a population, but need not be.
\item \code{S_obs}  Total number (not density) of all wild and hatchery-origin spawners.
\item \verb{n_age[min_age]_obs...n_age[max_age]_obs}  Multiple columns of
observed spawner age frequencies (i.e., counts), where \verb{[min_age]} and
\verb{[max_age]} are the numeral age in years (total, not ocean age) of the
youngest and oldest spawners, respectively. Complex age structures such as
Gilbert-Rich or maiden-kelt are not supported.
\item \code{n_W_obs}  Observed frequency of natural-origin spawners.
\item \code{n_H_obs}  Observed frequency of hatchery-origin spawners.
\item \code{F_rate}  Total harvest rate (proportion) of natural-origin fish.
\item \code{B_take_obs}  Number of adults taken for hatchery broodstock.
}}

\item{age_F}{Either a logical or 0/1 integer vector of length \code{N_age} indicating
whether each adult age is fully (non)selected by the fishery.
The default is all selected.}

\item{age_B}{Either a logical or 0/1 integer vector of length \code{N_age} indicating
whether each adult age is fully (non)selected in broodstock collection.
The default is all selected.}
}
\value{
A data frame with the following columns, some of which are simply
replicated from \code{fish_data}:
\itemize{
\item \code{pop}  See above.
\item \code{year}  See above.
\item \code{A}  See above.
\item \code{S_obs}  See above.
\item \verb{q_age[minAge]_obs...q_age[maxAge]_obs}  Multiple columns of spawner age
proportions corresponding to the frequencies in \code{fish_data}.
\item \code{p_age_minAge...p_age_maxAge}  Multiple columns of recruit age
proportions by brood year.
\item \code{p_HOS_obs}  Proportion of hatchery-origin spawners.
\item \code{F_rate}  See above.
\item \code{B_take_obs}  See above.
\item \code{R_obs}  Total natural-origin recruits from the brood year in each row.
}
}
\description{
Perform run reconstruction on brood table data
}
