% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonIPM-package.R
\docType{package}
\name{salmonIPM-package}
\alias{salmonIPM-package}
\alias{salmonIPM}
\title{The 'salmonIPM' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
