% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonIPM.R
\name{salmonIPM}
\alias{salmonIPM}
\title{Fits an integrated or run-reconstruction spawner-recruit model.}
\usage{
salmonIPM(
  model = "IPM",
  life_cycle = "SS",
  pool_pops = TRUE,
  SR_fun = "BH",
  stan_model = NULL,
  ages = NULL,
  par_models = NULL,
  scale = TRUE,
  fish_data,
  fish_data_fwd = NULL,
  fecundity_data = NULL,
  prior_data = NULL,
  age_S_obs = NULL,
  age_S_eff = NULL,
  conditionGRonMS = FALSE,
  init = NULL,
  pars = NULL,
  include = TRUE,
  log_lik = FALSE,
  chains,
  iter,
  warmup,
  thin = 1,
  cores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{model}{Either \code{"IPM"} or \code{"RR"}, indicating whether the data
are intended for an integrated or run-reconstruction model.}

\item{life_cycle}{Character string indicating which life-cycle model to fit.
One of the following options (must be \code{"SS"} if \code{model == "RR"}):
\itemize{
\item \code{"SS"}  Spawner-to-spawner (the default)
\item \code{"SSpa"}  Spawner-to-spawner with partially observed and partially
reproductively effective adult age classes
\item \code{"SMS"}  Spawner-smolt-spawner
\item \code{"SMaS"}  Spawner-smolt-spawner with multiple smolt age classes (currently only
available for \code{pool_pops == FALSE})
\item \code{"LCRchum"}  Customized spawner-egg-smolt-spawner model for Lower Columbia chum
(\code{pool_pops == TRUE})
\item \code{"ICchinook"}  Customized spawner-smolt-spawner model with downstream, SAR,
and upstream survival (\code{"pool_pops == TRUE"})
}}

\item{pool_pops}{Logical scalar defaulting to \code{TRUE}, indicating whether or not
to treat the different populations as hierarchical rather than
fixed/independent.}

\item{SR_fun}{One of \code{"exp"}, \code{"BH"} (the default), or \code{"Ricker"},
indicating which spawner-recruit function to fit.}

\item{stan_model}{Character string giving the name of the \strong{salmonIPM} model being
fit (".stan" filetype extension is not included). A more concise alternative to
specifying \code{model}, \code{life_cycle}, and \code{pool_pops} (and will override those arguments).}

\item{ages}{See \code{\link[=stan_data]{stan_data()}}.}

\item{par_models}{See \code{\link[=stan_data]{stan_data()}}.}

\item{scale}{See \code{\link[=stan_data]{stan_data()}}.}

\item{fish_data}{See \code{\link[=stan_data]{stan_data()}}.}

\item{fish_data_fwd}{See \code{\link[=stan_data]{stan_data()}}.}

\item{fecundity_data}{See \code{\link[=stan_data]{stan_data()}}.}

\item{prior_data}{See \code{\link[=stan_data]{stan_data()}}.}

\item{age_S_obs}{See \code{\link[=stan_data]{stan_data()}}.}

\item{age_S_eff}{See \code{\link[=stan_data]{stan_data()}}.}

\item{conditionGRonMS}{See \code{\link[=stan_data]{stan_data()}}.}

\item{init}{A list of named lists of initial values to be passed to
\code{\link[rstan:stan]{rstan::stan()}}. If \code{NULL}, initial values will be automatically
generated from the supplied data using \code{\link[=stan_init]{stan_init()}}.}

\item{pars}{Vector of character strings specifying parameters to monitor.
If NULL, default values are used. If a non-default value is supplied, the
user should make sure the parameters requested appear in the model configuration
specified.}

\item{include}{Logical scalar defaulting to \code{TRUE} indicating whether to include
or exclude the parameters given by \code{pars}. If \code{FALSE}, only entire
multidimensional parameters can be excluded, rather than particular elements of them.}

\item{log_lik}{Logical scalar indicating whether the pointwise log-likelihood
should be returned for later analysis with \code{\link[loo:loo]{loo::loo()}}.}

\item{chains}{Positive integer specifying the number of Markov chains.}

\item{iter}{Positive integer specifying the number of iterations for each
chain (including warmup).}

\item{warmup}{Positive integer specifying the number of warmup (aka burnin)
iterations per chain. If step-size adaptation is on (which it is by
default), this also controls the number of iterations for which adaptation
is run (and hence these warmup samples should not be used for inference).
The number of warmup iterations should not be larger than \code{iter}}

\item{thin}{Positive integer specifying the period for saving samples. The
default is 1, which is usually the recommended value.}

\item{cores}{Number of cores to use when executing the chains in parallel.
Defaults to one less than the number of cores available.}

\item{...}{Additional arguments to pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}.}
}
\value{
An object of class \code{stanfit} representing the fitted model. See
\code{\link[rstan:stan]{rstan::stan()}} for details.
}
\description{
Fits an integrated or run-reconstruction spawner-recruit model.
}
