% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonIPM.R
\name{salmonIPM}
\alias{salmonIPM}
\title{Fits an integrated or run-reconstruction spawner-recruit model.}
\usage{
salmonIPM(
  fish_data,
  fish_data_fwd = NULL,
  env_data = NULL,
  fecundity_data = NULL,
  prior_data = NULL,
  ages = NULL,
  age_S_obs = NULL,
  age_S_eff = NULL,
  conditionGRonMS = FALSE,
  model,
  life_cycle = "SS",
  pool_pops = TRUE,
  stan_model = NULL,
  SR_fun = "BH",
  init = NULL,
  pars = NULL,
  include = TRUE,
  log_lik = FALSE,
  chains,
  iter,
  warmup,
  thin = 1,
  cores = parallel::detectCores() - 1,
  ...
)
}
\arguments{
\item{fish_data}{See \code{\link[=stan_data]{stan_data()}}.}

\item{fish_data_fwd}{See \code{\link[=stan_data]{stan_data()}}.}

\item{env_data}{See \code{\link[=stan_data]{stan_data()}}.}

\item{fecundity_data}{See \code{\link[=stan_data]{stan_data()}}.}

\item{prior_data}{See \code{\link[=stan_data]{stan_data()}}.}

\item{ages}{See \code{\link[=stan_data]{stan_data()}}.}

\item{age_S_obs}{See \code{\link[=stan_data]{stan_data()}}.}

\item{age_S_eff}{See \code{\link[=stan_data]{stan_data()}}.}

\item{conditionGRonMS}{See \code{\link[=stan_data]{stan_data()}}.}

\item{model}{Either \code{"IPM"} or \code{"RR"}, indicating whether the data
are intended for an integrated or run-reconstruction model.}

\item{life_cycle}{Character string indicating which life-cycle model to fit.
Currently available options are spawner-to-spawner (\code{"SS"}, the
default), spawner-to-spawner "harvest model" (\code{"SS_F"}), or
spawner-smolt-spawner (\code{"SMS"}).}

\item{pool_pops}{Logical scalar defaulting to \code{TRUE}, indicating whether or not
to treat the different populations as hierarchical rather than
fixed/independent.}

\item{stan_model}{Character string giving the name of the \strong{salmonIPM} model being
fit (".stan" filetype extension is not included).}

\item{SR_fun}{One of \code{"exp"}, \code{"BH"} (the default), or
\code{"Ricker"}, indicating which spawner-recruit function to fit.}

\item{init}{A list of named lists of initial values to be passed to
\code{\link[rstan:stan]{rstan::stan()}}. If \code{NULL}, initial values will be automatically
generated from the supplied data using \code{\link[=stan_init]{stan_init()}}.}

\item{pars}{Vector of character strings specifying parameters to monitor.
If NULL, default values are used. If a non-default value is supplied, the
user should make sure the parameters requested appear in the model configuration
specified.}

\item{include}{Logical scalar defaulting to \code{TRUE} indicating whether to include
or exclude the parameters given by \code{pars}. If \code{FALSE}, only entire
multidimensional parameters can be excluded, rather than particular elements of them.}

\item{log_lik}{Logical scalar indicating whether the pointwise log-likelihood
should be returned for later analysis with \code{\link[loo:loo]{loo::loo()}}.}

\item{chains}{Positive integer specifying the number of Markov chains.}

\item{iter}{Positive integer specifying the number of iterations for each
chain (including warmup).}

\item{warmup}{Positive integer specifying the number of warmup (aka burnin)
iterations per chain. If step-size adaptation is on (which it is by
default), this also controls the number of iterations for which adaptation
is run (and hence these warmup samples should not be used for inference).
The number of warmup iterations should not be larger than \code{iter}}

\item{thin}{Positive integer specifying the period for saving samples. The
default is 1, which is usually the recommended value.}

\item{cores}{Number of cores to use when executing the chains in parallel.
Defaults to one less than the number of cores available.}

\item{...}{Additional arguments to pass to \code{\link[rstan:sampling]{rstan::sampling()}}.}
}
\value{
An object of class \code{stanfit} representing the fitted model. See
\code{\link[rstan:stan]{rstan::stan()}} for details.
}
\description{
Fits an integrated or run-reconstruction spawner-recruit model.
}
