% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR.R
\name{SR}
\alias{SR}
\title{Spawner-recruit functions}
\usage{
SR(SR_fun = "BH", alpha, Rmax, S, A = 1, R_per_S = FALSE)
}
\arguments{
\item{SR_fun}{One of \code{"exp"} (density-independent discrete exponential),
\code{"BH"} (Beverton-Holt, the default), or \code{"Ricker"}.
Synonyms \code{"B-H"}, \code{"bh"}, \code{"b-h"} and \code{"ricker"} are also accepted.}

\item{alpha}{A numeric vector, matrix or data frame of intrinsic productivity
(i.e., recruits per spawner at zero spawner density; slope of the spawner-recruit
function at the origin.).}

\item{Rmax}{A numeric vector, matrix or data frame of maximum recruitment per unit
of habitat (length or area). This corresponds to the asymptote of the Beverton-Holt
or the mode of the Ricker.}

\item{S}{A numeric vector, matrix or data frame of spawner abundance.}

\item{A}{A numeric vector, matrix or data frame of spawning habitat size
(either stream length or area), used to standardize \code{Rmax}. The default is 1,
in which case \code{Rmax} is in units of abundance rather than density.}

\item{R_per_S}{Logical indicating whether to return recruits per spawner rather than
recruits (the default).}
}
\description{
Compute recruitment given spawner abundance, a spawner-recruit function and parameters.
}
\details{
The \code{salmonIPM} package uses a nonstandard parameterization of the Ricker
model by the maximum recruitment \code{Rmax}. This is typically better identified by
data than the carrying capacity or per capita density dependence, and it
facilitates a common interpretation and priors with the Beverton-Holt.\tabular{rl}{
    \tab  \cr
   Discrete exponential: \tab \code{R = alpha * S} \cr
   Beverton-Holt: \tab \code{R = alpha * S / (1 + alpha * S / (A * Rmax))} \cr
   Ricker: \tab \code{R = alpha * S * exp(-alpha * S / (A * exp(1) * Rmax))} \cr
}


Calculations are vectorized and elements of shorter arguments are recycled
as necessary.
}
\examples{
alpha <- 3
Rmax <- 1000
S <- 500

SR(alpha = alpha, Rmax = Rmax, S = S) # default is Beverton-Holt
SR(alpha = alpha, Rmax = Rmax, S = 1e6) # approximately Rmax
SR(alpha = alpha, Rmax = Rmax, S = S, A = 0.1) # scale Rmax by habitat area
SR(SR_fun = "exp", alpha = alpha, Rmax = Rmax, S = S) # discrete exponential ignores Rmax
SR(alpha = rep(alpha, 10), Rmax = rep(Rmax, 10), S = matrix(S, 10, 4)) # vectorization with recycling
SR(alpha = alpha, Rmax = Rmax, S = S, R_per_S = TRUE) # return recruits per spawner

curve(SR(SR_fun = "Ricker", alpha = alpha, Rmax = Rmax, S = x), from = 0, to = 2000,
      xlab = "Spawners", ylab = "Recruits", main = "Ricker")
abline(h = Rmax, lty = 2)

}
\seealso{
\code{\link[=salmonIPM]{salmonIPM()}} for fitting models, \code{\link[=sim_salmonIPM]{sim_salmonIPM()}} for simulating data
}
