% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R, R/stan_data_v1.R
\name{stan_data}
\alias{stan_data}
\alias{stan_data}
\title{Assemble input data for integrated or run-reconstruction spawner-recruit models.}
\usage{
stan_data(fish_data, env_data = NULL, catch_data = NULL, model)

stan_data(fish_data, env_data = NULL, catch_data = NULL, model)
}
\arguments{
\item{fish_data}{Data frame that includes the following \code{colnames}, in no particular order except where noted:
\describe{
\item{\code{pop}}{Numeric or character population ID.}
\item{\code{year}}{Integer variable giving the year the fish spawned (i.e., the brood year).}
\item{\code{A}}{Spawning habitat size (either stream length or area). Will usually be time-invariant within a population, but need not be.}
\item{\code{S_tot_obs}}{Total number (not density) of wild and hatchery-origin spawners.}
\item{\code{n_age_minAge...n_age_maxAge}}{Multiple columns of observed spawner age frequencies (i.e., counts), where minAge (maxAge) is the numeral age in years (total, not ocean age) of the youngest (oldest) spawners.}
\item{\code{n_W_obs}}{Observed frequency of natural-origin spawners.}
\item{\code{n_H_obs}}{Observed frequency of hatchery-origin spawners.}
\item{\code{fit_p_HOS}}{Logical or 0/1 indicating for each row in fish_data whether the model should estimate p_HOS > 0. This is only required if model == "IPM".}
\item{\code{F_rate}}{Total harvest rate (proportion) of natural-origin fish, only if model != "IPM_F".}
\item{\code{B_take_obs}}{Number of adults taken for hatchery broodstock.}
}}

\item{env_data}{Optional data frame whose variables are time-varying environmental covariates, sequentially ordered with each row corresponding to a unique year in fish_data.}

\item{catch_data}{Only if model == "IPM_F", a data frame with numeric columns
\describe{
\item{\code{year}}{Year for fishery data. Must be identical to \code{unique(fish_data$year)}.}
\item{\code{R_F_obs}}{Total recruits to the fishery.}
\item{\code{C_obs}}{Total catch.}
}}

\item{model}{One of \code{"IPM"}, \code{"RR"}, or \code{"IPM_F"}, indicating whether the data are intended for an integrated or run-reconstruction model or the integrated "harvest" model.}

\item{fish_data_fwd}{Only if model == "IPM", optional data frame with the following \code{colnames}, representing "forward" or "future" simulations:
\describe{
\item{\code{pop}}{Numeric or character population ID. All values must also appear in \code{fish_data$pop}.}
\item{\code{year}}{Integer variable giving the year the fish spawned (i.e., the brood year). For each population in \code{fish_data_fwd$pop}, the first year appearing in \code{fish_data_fwd$year} must be one greater than the last year appearing in \code{fish_data$year}, i.e., \code{min(fish_data_fwd$year[fish_data_fwd$pop==j]) == max(fish_data$year[fish_data$pop==j]) + 1}.}
\item{\code{A}}{Spawning habitat size (either stream length or area). Will usually be time-invariant within a population, but need not be.}
\item{\code{F_rate}}{Total harvest rate (proportion) of natural-origin fish.}
\item{\code{B_rate}}{Total broodstock removal rate (proportion) of natural-origin fish.}
\item{\code{p_HOS}}{Proportion of hatchery-origin spawners.}
}}

\item{fish_data}{Data frame that includes the following \code{colnames}, in no particular order except where noted:
\describe{
\item{\code{pop}}{Numeric or character population ID.}
\item{\code{year}}{Numeric variable giving the year the fish spawned (i.e., the brood year).}
\item{\code{A}}{Spawning habitat size (either stream length or area). Will usually be time-invariant within a population, but need not be.}
\item{\code{S_tot_obs}}{Total number (not density) of wild and hatchery-origin spawners.}
\item{\code{n_age_minAge...n_age_maxAge}}{Multiple columns of observed spawner age frequencies (i.e., counts), where minAge (maxAge) is the numeral age in years (total, not ocean age) of the youngest (oldest) spawners.}
\item{\code{n_W_obs}}{Observed frequency of natural-origin spawners.}
\item{\code{n_H_obs}}{Observed frequency of hatchery-origin spawners.}
\item{\code{fit_p_HOS}}{Logical or 0/1 indicating for each row in fish_data whether the model should estimate p_HOS > 0. This is only required if model == "IPM".}
\item{\code{F_rate}}{Total harvest rate (proportion) of natural-origin fish, only if model != "IPM_F".}
\item{\code{B_take_obs}}{Number of adults taken for hatchery broodstock.}
}}

\item{catch_data}{Only if model == "IPM_F", a data frame with numeric columns
\describe{
\item{\code{year}}{Year for fishery data. Must be identical to \code{unique(fish_data$year)}.}
\item{\code{R_F_obs}}{Total recruits to the fishery.}
\item{\code{C_obs}}{Total catch.}
}}

\item{model}{One of \code{"IPM"}, \code{"RR"}, or \code{"IPM_F"}, indicating whether the data are intended for an integrated or run-reconstruction model or the integrated "harvest" model.}
}
\value{
A named list that can be passed to \code{stan} as the \code{data} argument.

A named list that can be passed to \code{stan} as the \code{data} argument.
}
\description{
Assemble input data for integrated or run-reconstruction spawner-recruit models.

Assembles input data in a format that can be passed to Stan for fitting either integrated or run-reconstruction spawner-recruit models.
}
