% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{stan_data}
\alias{stan_data}
\title{Assemble input data for integrated or run-reconstruction spawner-recruit
models.}
\usage{
stan_data(fish_data, fish_data_fwd = NULL, env_data = NULL,
  ages = NULL, age_S_obs = NULL, age_S_eff = NULL, stan_model,
  SR_fun = "BH")
}
\arguments{
\item{fish_data}{Data frame that includes the following \code{colnames}, in
no particular order except where noted: \describe{
\item{\code{pop}}{Numeric or character population ID.}
\item{\code{year}}{Numeric variable giving the year the fish spawned (i.e.,
the brood year).} \item{\code{A}}{Spawning habitat size (either stream
length or area). Will usually be time-invariant within a population, but
need not be.} \item{\code{M_obs}}{Total number of wild-origin smolts (only
needed for models including smolt stage).}
\item{\code{n_Mage[min_Mage]_obs...n_Mage[max_Mage]_obs}}{If
\code{life_cycle=="SMaS"}, multiple columns of observed smolt age
frequencies (i.e., counts), where \code{[min_Mage]} and \code{[max_Mage]}
are the numeral age in years of the youngest and oldest smolts,
respectively. Note that age is measured in calendar years from the brood
year (i.e., the Gilbert-Rich system).} \item{\code{S_obs}}{Total number
(not density) of wild and hatchery-origin spawners.}
\item{\code{n_age[min_age]_obs...n_age[max_age]_obs}}{Multiple columns of
observed spawner age frequencies (i.e., counts), where \code{[min_age]} and
\code{[max_age]} are the numeral age in years (total, not ocean age) of the
youngest and oldest spawners, respectively.}
\item{\code{n_MSage[min_MSage]_obs...n_MSage[max_MSage]_obs}}{If
\code{life_cycle=="SMaS"}, multiple columns of observed ocean age
frequencies (i.e., counts), where \code{[min_MSage]} and \code{[max_MSage]}
are the youngest and oldest ocean age in years, respectively.}
\item{\code{n_GRage[min_age]_[min_Mage]_obs...n_GRage[max_age]_[max_Mage]_obs}}{If
 \code{life_cycle=="SMaS"}, multiple columns of observed Gilbert-Rich age
frequencies, sorted first by smolt age (\code{min_Mage:max_Mage}) and then
by total age \code{min_age:max_age}. For example, a life history with
subyearling or yearling smolts and ocean ages 2:3 would have column names
c("n_GRage_3_1_obs", "n_GRage_4_1_obs", "n_GRage_4_2_obs",
"n_GRage_5_2_obs")} \item{\code{n_W_obs}}{Observed frequency of
natural-origin spawners.} \item{\code{n_H_obs}}{Observed frequency of
hatchery-origin spawners.} \item{\code{fit_p_HOS}}{Logical or 0/1
indicating for each row in fish_data whether the model should estimate
\code{p_HOS > 0}. This is only required if \code{model == "IPM"}.}
\item{\code{F_rate}}{Total harvest rate (proportion) of natural-origin
fish.} \item{\code{B_take_obs}}{Number of adults taken for hatchery
broodstock.} }}

\item{fish_data_fwd}{Only if `stan_model == "IPM_SS_pp"`, optional data frame
with the following \code{colnames}, representing "forward" or "future"
simulations. Unlike \code{fish_data}, a given combination of population and
year may occur multiple times, perhaps to facilitate comparisons across
scenarios or "branches" with different inputs (e.g., harvest rate). In this
case, all branches are subjected to the same sequence of process errors in
recruitment and age structure. \describe{ \item{\code{pop}}{Numeric or
character population ID. All values must also appear in
\code{fish_data$pop}.} \item{\code{year}}{Integer variable giving the year
the fish spawned (i.e., the brood year). For each population in
\code{fish_data_fwd$pop}, the first year appearing in
\code{fish_data_fwd$year} must be one greater than the last year appearing
in \code{fish_data$year}, i.e.,
\code{min(fish_data_fwd$year[fish_data_fwd$pop==j]) ==
max(fish_data$year[fish_data$pop==j]) + 1}.} \item{\code{A}}{Spawning
habitat size (either stream length or area). Will usually be time-invariant
within a population, but need not be.} \item{\code{F_rate}}{Total harvest
rate (proportion) of natural-origin fish.} \item{\code{B_rate}}{Total
broodstock removal rate (proportion) of natural-origin fish.}
\item{\code{p_HOS}}{Proportion of hatchery-origin spawners.} }}

\item{env_data}{Optional data frame or named list of data frames whose
variables are time-varying environmental covariates, sequentially ordered
with each row corresponding to a unique year in fish_data. If a named list,
element names correspond to stage- or transition-specific covariate
matrices defined in the Stan model being used. (This is required for
multi-stage models.)}

\item{ages}{For multi-stage models, a named list giving the fixed ages in
years of all subadult life stages.}

\item{age_S_obs}{Only if `stan_model == "IPM_SSpa_np"`, a logical or numeric
vector indicating, for each adult age, whether observed total spawner data
includes that age. The default is to treat `S_obs` as including spawners of
all ages.}

\item{age_S_eff}{Only if `stan_model == "IPM_SSpa_np"`, a logical or numeric
vector indicating, for each adult age, whether spawners of that age
contribute toward reproduction. This could be used, e.g., to exclude jacks
from the effective breeding population. The default is to include spawners
of all ages.}

\item{stan_model}{Character string giving the name of the Stan model being
fit (".stan" filetype extension is not included).}

\item{SR_fun}{One of \code{"exp"}, \code{"BH"} (the default), or
\code{"Ricker"}, indicating which spawner-recruit function to fit.}
}
\value{
A named list that can be passed to \code{stan} as the \code{data}
  argument.
}
\description{
Assemble input data for integrated or run-reconstruction spawner-recruit
models.
}
