% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_init.R
\name{stan_init}
\alias{stan_init}
\title{Generate initial values for fitting either integrated or run-reconstruction
spawner-recruit models in Stan.}
\usage{
stan_init(data, stan_model, chains)
}
\arguments{
\item{data}{Named list of input data for fitting either an integrated or
run-reconstruction spawner-recruit model in Stan, as returned by
\code{stan_data}.}

\item{stan_model}{Character string giving the name of the Stan model being
fit (".stan" filetype extension is not included).}

\item{chains}{A positive integer specifying the number of Markov chains.}
}
\value{
A list with initial starting values for the parameters and states in
  the Stan model.
}
\description{
Generate initial values for fitting either integrated or run-reconstruction
spawner-recruit models in Stan.
}
