% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_pars.R
\name{stan_pars}
\alias{stan_pars}
\title{Fits an integrated or run-reconstruction spawner-recruit model.}
\usage{
stan_pars(stan_model)
}
\arguments{
\item{stan_model}{Character string giving the name of the Stan model being
fit (".stan" filetype extension is not included).}
}
\description{
Fits an integrated or run-reconstruction spawner-recruit model.
}
