% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_pars.R
\name{stan_pars}
\alias{stan_pars}
\title{Return the (hyper)parameters and states in a specified \strong{salmonIPM} model}
\usage{
stan_pars(
  stan_model = c("IPM_SS_np", "IPM_SSiter_np", "IPM_SS_pp", "IPM_SSiter_pp",
    "IPM_SMS_np", "IPM_SMS_pp", "IPM_SMaS_np", "IPM_LCRchum_pp", "IPM_ICchinook_pp",
    "RR_SS_np", "RR_SS_pp"),
  pars = c("all", "hyper", "group", "states", "ppd")
)
}
\arguments{
\item{stan_model}{Character string giving the name of the model.
See \code{\link[=salmonIPM]{salmonIPM()}} for details.}

\item{pars}{An optional character vector specifying one or more hierarchical levels
of parameters. Options are \code{"all"} (the default), \code{"hyper"} (top-level parameters that
are given priors), \code{"group"} (\code{pop}- or \code{year}-level parameters shared by multiple states),
\code{"states"} (the lowest level, corresponding to unique rows in \code{fish_data}), and
\code{"ppd"} (only if \code{model == "RR"}, observation-level predictions drawn from the posterior
predictive distribution).

Using pars="beta" will restrict the displayed parameters to only the regression coefficients (without the intercept). "alpha" can also be used as a shortcut for "(Intercept)". If the model has varying intercepts and/or slopes they can be selected using pars = "varying".}
}
\value{
Character vector with names of selected parameters and states
}
\description{
Return the (hyper)parameters and states in a specified \strong{salmonIPM} model
}
